[Decomp][GameObject] Cohesion split: extract Update loop to GameObjectUpdate.cpp
